# go/multi_class_classification_result.go
/*
 * Predict
 *
 * Endpoints API for classification and other prediction services in SKIL
 *
 * API version: 1.0.3
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package skilclient

// Represents all of the labels for a given classification
type MultiClassClassificationResult struct {

	RankedOutcomes [][]string `json:"rankedOutcomes,omitempty"`

	MaxOutcomes []string `json:"maxOutcomes,omitempty"`

	Probabilities [][]float32 `json:"probabilities,omitempty"`
}