# scala/src/main/scala/ai/skymind/skil/model/Prediction.scala
/**
 * Predict
 * Endpoints API for classification and other prediction services in SKIL
 *
 * OpenAPI spec version: 1.0.3
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

package ai.skymind.skil.model


case class Prediction (
  Id: Option[String] = None,
  NeedsPreProcessing: Option[Boolean] = None,
  Prediction: Option[INDArray] = None
)